refactor(common-types): relocate cross-channel wire mapping + fallback environment
mapCrossChannelToApiFormat and buildFallbackEnvironment move to
crossChannelEnvironment.ts so the worker-side assembler serializes
cross-channel groups through the SAME functions the bot path uses — wire
parity by construction. What stays caller-side is HOW a channel id resolves
to an environment: the bot live-fetches from Discord, the worker reads the
envelope's knownChannelEnvironments map; both degrade to the shared
fallback. The mapper's tests move with it; bot consumers import from the
package directly (no re-export wrapper).

Scrub-pass fold-forwards riding along: the envelope builder's refs param
renamed authorDisplayName → rawAuthorDisplayName (consistency with its
raw* siblings), the relocated persona resolver's module-doc sentence
fragment repaired, and a shouty test name lowercased.
