+++ packages/common-types/src/index.ts
@@ -92,6 +92,7 @@ export * from './utils/extendedContextPersonaResolver.js';
 export * from './utils/referenceEnrichment.js';
 export * from './utils/messageLinkParser.js';
 export * from './utils/mentionRewriter.js';
+export * from './utils/crossChannelEnvironment.js';
 export * from './services/UserService.js';
 export * from './services/BaseCacheInvalidationService.js';
 export * from './services/CacheInvalidationService.js';
